requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
Pillow==10.0.1
django-extensions==3.2.3
dj-database-url==2.1.0
bleach==6.1.0
mysqlclient==2.2.0